Level1 first version
Still need to fix some issue:
Use numberFormatter for the currency display.

+++ src/main/java/model/Transaction.java
@@ -53,7 +53,7 @@ public class Transaction {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 }
